Addressed Changes
