+++ lom/src/plugins/sonic/plugin_integration_tests/linkcrc_mocker/main.go
@@ -1,29 +1,14 @@
 package main
 
 import (
-    "github.com/go-redis/redis"
-    "lom/src/plugins/sonic/plugin_integration_tests/utils"
     "os"
     "os/exec"
-    "strconv"
-    "time"
+    "lom/src/plugins/sonic/plugin_integration_tests/linkcrc_mocker/linkcrc_utils"
+    "lom/src/plugins/sonic/plugin_integration_tests/utils"
 )
 
 const (
-    redis_address                    = "localhost:6379"
-    redis_counters_db                = 2
-    redis_app_db                     = 0
-    fileName                         = "./LINK_CRC_COUNTERS"
-    redis_password                   = ""
-    counter_poll_disable_command     = "sudo counterpoll port disable"
-    action_name                      = "link_crc_detection"
-    detection_type                   = "detection"
-    counter_poll_enable_command      = "sudo counterpoll port enable"
-    counters_port_name_map_redis_key = "COUNTERS_PORT_NAME_MAP"
-    admin_status                     = "admin_status"
-    oper_status                      = "oper_status"
-    ifUp                             = "up"
-    counters_db                      = "COUNTERS:"
+counter_poll_disable_command = "sudo counterpoll port disable"
 )
 
 func main() {
@@ -35,91 +20,26 @@ func main() {
     } else {
         utils.PrintInfo("Successfuly Disabled counterpoll")
     }
-    MockRedisWithLinkCrcCounters()
+
+	testId := os.Args[1]
+	interfaces := os.Args[2:]
+
+	switch testId {
+	case "0":
+		linkcrc_utils.MockRedisWithLinkCrcCounters("1", 10, interfaces)
+		break
+	case "1":
+		linkcrc_utils.MockRedisWithLinkCrcCounters("0", 10, interfaces)
+		break
+	case "2":
+		linkcrc_utils.MockRedisWithLinkCrcCounters("1,0,0,0", 10, interfaces)
+		break
+	case "3":
+		linkcrc_utils.MockRedisWithLinkCrcCounters("1,0,0,0,0", 2, interfaces)
+		break
+	default:
+		utils.PrintError("Invalid test Id %d", testId)
+	}
+
 }
 
-func MockRedisWithLinkCrcCounters() {
-    var countersDbClient = redis.NewClient(&redis.Options{
-        Addr:     redis_address,
-        Password: redis_password,
-        DB:       redis_counters_db,
-    })
-
-    // Get all ifName to oid mappings.
-    interfaceToOidMapping, err := countersDbClient.HGetAll(counters_port_name_map_redis_key).Result()
-    if err != nil {
-        utils.PrintError("Error fetching counters port name map %v", err)
-        return
-    }
-
-    // Filter only those required as per the links passed through arguments
-    mockedLinks := make(map[string]string)
-    for _, v := range os.Args[3:] {
-        mockedLinks[v] = interfaceToOidMapping[v]
-    }
-
-    // Set admin and oper status as up for mocked links.
-    var appDbClient = redis.NewClient(&redis.Options{
-        Addr:     redis_address,
-        Password: redis_password,
-        DB:       redis_app_db,
-    })
-    ifStatusMock := map[string]interface{}{admin_status: ifUp, oper_status: ifUp}
-    for k, _ := range mockedLinks {
-        _, err = appDbClient.HMSet("PORT_TABLE:"+k, ifStatusMock).Result()
-        if err != nil {
-            utils.PrintError("Error mocking admin and oper status for interface %s. Err: %v", k, err)
-            return
-        } else {
-            utils.PrintInfo("Successfully Mocked admin and oper status for interface %s", k)
-        }
-    }
-
-    var ifInErrors float64
-    var ifInUnicastPackets float64
-    var ifOutUnicastPackets float64
-    var ifOutErrors float64
-    ifInErrors = 100
-    ifInUnicastPackets = 101
-    ifOutUnicastPackets = 1100
-    ifOutErrors = 1
-
-    period, err := strconv.Atoi(os.Args[2])
-    timer1 := time.NewTimer(time.Duration(period) * time.Minute)
-
-    simulate, err := strconv.Atoi(os.Args[1]) /* 0 indicates simulation, 1 indicates non-simulation with non-zero counters */
-
-loop:
-    for {
-        select {
-        case <-timer1.C:
-            break loop
-        default:
-
-            datapoint := map[string]interface{}{"SAI_PORT_STAT_IF_IN_ERRORS": ifInErrors, "SAI_PORT_STAT_IF_IN_UCAST_PKTS": ifInUnicastPackets, "SAI_PORT_STAT_IF_OUT_UCAST_PKTS": ifOutUnicastPackets, "SAI_PORT_STAT_IF_OUT_ERRORS": ifOutErrors}
-            for ifName, oidMapping := range mockedLinks {
-                _, err := countersDbClient.HMSet(counters_db+oidMapping, datapoint).Result()
-                if err != nil {
-                    utils.PrintError("Error mocking redis data for interface %s. Err %v", ifName, err)
-                    return
-                } else {
-                    utils.PrintInfo("Successfuly mocked redis data for interface %s", ifName)
-                }
-            }
-            time.Sleep(2 * time.Second)
-
-            if simulate == 0 {
-                ifInErrors = ifInErrors + 40
-                ifInUnicastPackets = ifInUnicastPackets + 67
-                ifOutUnicastPackets = ifOutUnicastPackets + 67
-                ifOutErrors = ifOutErrors + 2
-            } else {
-                ifInErrors = ifInErrors + 14
-                ifInUnicastPackets = ifInUnicastPackets + 20000001
-                ifOutUnicastPackets = ifOutUnicastPackets + 67
-                ifOutErrors = ifOutErrors + 2
-            }
-
-        }
-    }
-}
